requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
ollama==0.4.4
chromadb==0.5.23
Pillow==11.0.0
piexif==1.1.3
pydantic==2.10.3
python-multipart==0.0.19
